src/main/java/com/example/springmockito/controller/MovieController.java
package com.example.springmockito.controller;

import com.example.springmockito.model.Movie;
import com.example.springmockito.service.MovieService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import reactor.core.publisher.Flux;

import java.util.List;

@RestController
@RequestMapping("/api")
public class MovieController {

    private final MovieService movieService;

    public MovieController(MovieService movieService) {
        this.movieService = movieService;
    }

    @GetMapping("/movies")
    public Flux<Movie> getMovies() {
        return movieService.getMovies();
    }
}
